style: Center quote box and update container alignment
Align the quote box in the center of the page by adding a new CSS class called "quotebox" and wrapping the quote element with it. Also, update the alignment of the container to center its content both vertically and horizontally by adding "align-items: center" and "justify-content: center" properties to the CSS rules for the container class.